# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  common.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  dto.LoginRequest:
    properties:
      captchaToken:
        description: |-
          CaptchaToken /captcha/verify 校验通过后返回的一次性凭证。
          必填：留空意味着攻击者可直接 POST /auth/login 绕过人机校验。
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - captchaToken
    - password
    - username
    type: object
  dto.RefreshTokenRequest:
    properties:
      refreshToken:
        description: |-
          RefreshToken 待刷新的 refresh token。

          **刻意不是 `binding:"required"`**（P3-B2 起）：浏览器路径下它由 HttpOnly
          cookie 携带，而前端 JS 读不到 HttpOnly cookie，也就**不可能**把它放进请求体 ——
          请求体是空的。真正「有没有凭据」的判定在 Controller 里用
          authcookie.RefreshFromRequest 统一做（请求体优先、其次 cookie，
          两者都取不到才拒绝）。

          保留该字段是为了不破坏 Swagger / curl / 第三方这类显式传参的调用方。
          若哪天有人"顺手"把 required 加回来，现象是「浏览器端 token 一过期就被
          踢到登录页」—— 与 B4 要修的那个缺陷一模一样，极难怀疑到参数绑定这一层。
        type: string
    type: object
...
  vo.LoginResponse:
    properties:
      accessToken:
        type: string
      expiresIn:
        type: integer
      refreshToken:
        type: string
      tokenType:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Gin-Admin 后台管理系统 API 文档
  title: Gin-Admin API
  version: 1.0.0
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 登录参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/vo.LoginResponse'
              type: object
      summary: 用户登录
      tags:
      - 认证
  /auth/logout:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 退出登录
      tags:
      - 认证
  /auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: RefreshToken（浏览器路径下可省略，凭据在 HttpOnly cookie 里）
        in: body
        name: body
        schema:
          $ref: '#/definitions/dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/vo.LoginResponse'
              type: object
      summary: 刷新Token
      tags:
      - 认证
  /auth/userInfo:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/vo.UserInfoResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 获取用户信息
      tags:
      - 认证
...
  /captcha/verify:
    post:
      consumes:
      - application/json
      parameters:
      - description: 验证码 token 与点击坐标
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.CaptchaVerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      summary: 校验点选验证码
      tags:
      - 验证码
...
  /member:
    post:
      consumes:
      - application/json
      parameters:
      - description: 会员信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateMemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 创建会员
      tags:
      - 会员
    put:
      consumes:
      - application/json
      parameters:
      - description: 会员信息（含 id）
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateMemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 更新会员
      tags:
      - 会员
  /member/{id}:
    delete:
      parameters:
      - description: 会员 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 删除会员
      tags:
      - 会员
    get:
      parameters:
      - description: 会员 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 会员详情
      tags:
      - 会员
  /member/level:
    post:
      consumes:
      - application/json
      parameters:
      - description: 等级信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateMemberLevelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 创建会员等级
      tags:
      - 会员等级
    put:
      consumes:
      - application/json
      parameters:
      - description: 等级信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateMemberLevelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 更新会员等级
      tags:
      - 会员等级
  /member/level/{id}:
    delete:
      parameters:
      - description: 等级ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 删除会员等级
      tags:
      - 会员等级
  /member/level/all:
    get:
      description: 给下拉选项用；`/member/level/list` 是分页版
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 全部会员等级（不分页）
      tags:
      - 会员等级
  /member/level/list:
    get:
      parameters:
      - description: 等级名称
        in: query
        name: name
        type: string
      - description: 页码
        in: query
        name: page
        required: true
        type: integer
      - description: 每页条数
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 等级列表
      tags:
      - 会员等级
  /member/list:
    get:
      parameters:
      - description: 手机号
        in: query
        name: phone
        type: string
      - description: 昵称（模糊）
        in: query
        name: nickname
        type: string
      - description: 等级 ID
        in: query
        name: levelId
        type: integer
      - description: 状态：0 停用 / 1 正常
        in: query
        name: status
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页条数
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 会员列表
      tags:
      - 会员
  /member/points/list:
    get:
      parameters:
      - description: 会员ID
        in: query
        name: memberId
        type: integer
      - description: 类型 1获取 2消费
        in: query
        name: type
        type: integer
      - description: 页码
        in: query
        name: page
        required: true
        type: integer
      - description: 每页条数
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 积分明细列表
      tags:
      - 积分明细
  /member/status:
    put:
      consumes:
      - application/json
      parameters:
      - description: 会员 ID 与目标状态
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateMemberStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 启用 / 停用会员
      tags:
      - 会员
  /member/tag:
    post:
      consumes:
      - application/json
      parameters:
      - description: 标签信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateMemberTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 创建会员标签
      tags:
      - 会员标签
    put:
      consumes:
      - application/json
      parameters:
      - description: 标签信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateMemberTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 更新会员标签
      tags:
      - 会员标签
  /member/tag/{id}:
    delete:
      parameters:
      - description: 标签ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 删除会员标签
      tags:
      - 会员标签
  /member/tag/all:
    get:
      description: 给下拉选项用；`/member/tag/list` 是分页版
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 全部会员标签（不分页）
      tags:
      - 会员标签
  /member/tag/list:
    get:
      parameters:
      - description: 标签名称
        in: query
        name: name
        type: string
      - description: 页码
        in: query
        name: page
        required: true
        type: integer
      - description: 每页条数
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 标签列表
      tags:
      - 会员标签
  /member/tags:
    put:
      consumes:
      - application/json
      parameters:
      - description: 会员 ID 与标签 ID 列表
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateMemberTagsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 设置会员标签
      tags:
      - 会员
  /member/visit:
    put:
      consumes:
      - application/json
      parameters:
      - description: 会员 ID
        in: body
        name: id
        required: true
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 更新最后访问时间
      tags:
      - 会员
...
  /system/agreement:
    post:
      consumes:
      - application/json
      parameters:
      - description: 标题
        in: body
        name: title
        required: true
        schema:
          type: string
      - description: 正文
        in: body
        name: content
        schema:
          type: string
      - description: 协议类型
        in: body
        name: type
        required: true
        schema:
          type: string
      - description: 排序
        in: body
        name: sort
        schema:
          type: integer
      - description: 状态：0 停用 / 1 正常
        in: body
        name: status
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 创建协议
      tags:
      - 协议
    put:
      consumes:
      - application/json
      parameters:
      - description: 协议 ID
        in: body
        name: id
        required: true
        schema:
          type: integer
      - description: 标题
        in: body
        name: title
        required: true
        schema:
          type: string
      - description: 正文
        in: body
        name: content
        schema:
          type: string
      - description: 协议类型
        in: body
        name: type
        required: true
        schema:
          type: string
      - description: 排序
        in: body
        name: sort
        schema:
          type: integer
      - description: 状态：0 停用 / 1 正常
        in: body
        name: status
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 更新协议
      tags:
      - 协议
  /system/agreement/{id}:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 删除协议
      tags:
      - 协议
  /system/agreement/list:
    get:
      parameters:
      - description: 标题（模糊）
        in: query
        name: name
        type: string
      - description: 协议类型
        in: query
        name: type
        type: string
      - description: 状态
        in: query
        name: status
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页条数
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 协议列表
      tags:
      - 协议
  /system/agreement/type/{type}:
    get:
      description: 前端协议页用；同类型取排序最靠前的一条
      parameters:
      - description: 协议类型
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 按类型取协议
      tags:
      - 协议
...
  /system/config/{id}:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 删除参数
      tags:
      - 参数配置
  /system/config/batch:
    put:
      consumes:
      - application/json
      description: 按前缀成组保存（如站点设置表单一次提交多项）
      parameters:
      - description: 键名前缀
        in: body
        name: prefix
        required: true
        schema:
          type: string
      - description: 键值对列表，元素形如 {key, value}
        in: body
        name: items
        required: true
        schema:
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 批量保存参数
      tags:
      - 参数配置
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: '输入格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"